team-cost-reporter/requirements.txt
pyyaml
jsonmerge
